requirements.txt
amqp==5.1.1
arabic-reshaper==3.0.0
asgiref==3.7.1
asn1crypto==1.5.1
async-timeout==4.0.2
billiard==4.1.0
Brotli==1.0.9
celery==5.3.1
certifi==2023.5.7
cffi==1.15.1
chardet==5.1.0
charset-normalizer==3.2.0
click==8.1.4
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
cryptography==41.0.2
cssselect2==0.7.0
Django==4.2.1
django-bootstrap5==23.1
django-cleanup==7.0.0
django-recaptcha==3.0.0
easy-thumbnails==2.8.5
et-xmlfile==1.1.0
fonttools==4.41.1
html5lib==1.1
idna==3.4
jdcal==1.4.1
kombu==5.3.1
lml==0.1.0
lxml==4.9.3
openpyxl==3.0.0
oscrypto==1.3.0
Pillow==9.5.0
prompt-toolkit==3.0.39
psycopg2==2.9.6
pycparser==2.21
pydyf==0.7.0
pyexcel==0.7.0
pyexcel-io==0.6.6
pyexcel-xls==0.7.0
pyexcel-xlsx==0.6.0
pyHanko==0.19.0
pyhanko-certvalidator==0.23.0
pypdf==3.12.1
pyphen==0.14.0
pypng==0.20220715.0
python-bidi==0.4.2
python-dateutil==2.8.2
# PyYAML==6.0
wheel==0.40.0
qrcode==7.4.2
redis==4.6.0
# reportlab==3.6.13
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
static-thumbnails==0.9
# svglib==1.5.1
texttable==1.6.7
tinycss2==1.2.1
typing_extensions==4.6.1
tzdata==2023.3
tzlocal==5.0.1
uritools==4.0.1
urllib3==2.0.3
vine==5.0.0
wcwidth==0.2.6
weasyprint==59.0
webencodings==0.5.1
# xhtml2pdf==0.2.11
xlrd==2.0.1
xlwings==0.30.10
xlwt==1.3.0
zopfli==0.2.2
gunicorn==20.1.0
